% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{CompareBirdSpeciesBarplot}
\alias{CompareBirdSpeciesBarplot}
\title{Draws a Barplot of a Bird Species}
\usage{
CompareBirdSpeciesBarplot(bird1, bird2, main.title = "", generate.svg = F,
  cbc.code = "_CBC", cbc.name = "")
}
\arguments{
\item{main.title}{A title to put on the graph; if not provided, species name used}

\item{generate.svg}{Generates an SVG image}

\item{cbc.code}{The code for the CBC}

\item{bird}{A bird species list object}
}
\description{
\code{DrawBirdSpeciesBarplot} will create a barplot graph of a bird species.
}

