% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DrawBirdSpeciesBarplot}
\alias{DrawBirdSpeciesBarplot}
\title{Draws a Barplot of a Bird Species}
\usage{
DrawBirdSpeciesBarplot(bird, main.title = "", generate.svg = F,
  cbc.code = "_CBC", cbc.name = "")
}
\arguments{
\item{bird}{A bird species list object}

\item{main.title}{A title to put on the graph; if not provided, species name used}

\item{generate.svg}{Generates an SVG image}

\item{cbc.code}{The code for the CBC}
}
\description{
\code{DrawBirdSpeciesBarplot} will create a barplot graph of a bird species.
}

