% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{FileSafeName}
\alias{FileSafeName}
\title{Converts a species or taxon name to be filename-safe}
\usage{
FileSafeName(sx)
}
\arguments{
\item{sx}{A string}
}
\value{
A string
}
\description{
\code{FileSafeName} will take a string and return one that is safe as a filename
}

