% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GenerateAllSvgs}
\alias{GenerateAllSvgs}
\title{Generates Barplot SVGs for All Taxons in a Count}
\usage{
GenerateAllSvgs(cbc)
}
\arguments{
\item{\code{cbc}}{The CBC count vector}
}
\description{
\code{GenerateAllSvgs} will create SVGs for all taxons.
}

