% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ImportCbcCsv}
\alias{ImportCbcCsv}
\title{Import a Normalized CBC CSV File}
\usage{
ImportCbcCsv(csv.file)
}
\arguments{
\item{file}{A CSV file}
}
\value{
a \code{cbc.count} object
}
\description{
\code{ImportCbcCsv} takes a file generated by the CBC-CSV-CLI (C3) tool.
}

