% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{MakeSvgDirectory}
\alias{MakeSvgDirectory}
\title{Makes a directory for SVGs of a CBC, if it does not already exist}
\usage{
MakeSvgDirectory(sx)
}
\arguments{
\item{sx}{A string}
}
\description{
\code{MakeSvgDirectory} will create a sub-directory if it doesn't exist yet in the ./svgs folder
}

