% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{PlotAreaChart}
\alias{PlotAreaChart}
\title{Plot a Google Charts Area Chart}
\usage{
PlotAreaChart(cbc, taxons = c("Canada Goose"), embed.file = F,
  cbc.code = "_CBC", chart.width = NA)
}
\arguments{
\item{cbc}{The cbc dataframe}

\item{taxons}{A collection of taxon names to include}

\item{chart.width}{Width of the chart}
}
\description{
\code{PlotAreaChart} will create an area chart of a bird species.
}

