% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{RemoveXHeaders}
\alias{RemoveXHeaders}
\title{Removes the 'X' prefix from column names in data frames that are years}
\usage{
RemoveXHeaders(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
\code{RemoveXHeaders} will remove the prefix 'X' from a data frame's headers
}

