% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SelectSpecies}
\alias{SelectSpecies}
\title{Select a Bird Species}
\usage{
SelectSpecies(cbc, species.name)
}
\arguments{
\item{cbc}{A \code{cbc} list object}

\item{species.name}{A bird species or taxon name string}
}
\value{
a \code{bird} list object
}
\description{
\code{SelectSpecies} takes a \code{cbc.count} object and a \code{species.name}
  string and returns a \code{bird} object.
}

