% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ImportCbcCsv}
\alias{ImportCbcCsv}
\title{Import a Normalized CBC CSV File}
\usage{
ImportCbcCsv(csv.file, cbc.name = "")
}
\arguments{
\item{csv.file}{A CSV file}

\item{cbc.name}{Optional. The name of the CBC}
}
\value{
a \code{cbc} list object
}
\description{
\code{ImportCbcCsv} takes a file generated by the CBC-CSV-CLI (C3) tool.
}

