% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SelectSpecies}
\alias{SelectSpecies}
\title{Select a Bird Species}
\usage{
SelectSpecies(cbc.count, species.name)
}
\arguments{
\item{cbc.count}{A \code{cbc.count} object}

\item{species.name}{A bird species name string}
}
\value{
a \code{bird.species} object
}
\description{
\code{SelectSpecies} takes a \code{cbc.count} object and a \code{species.name}
  string and returns a \code{bird} object.
}

